etl/requirements.txt
elasticsearch==7.17.4
psycopg2-binary==2.9.3
python-dotenv==0.20.0
